requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==22.1.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
Flask==2.0.2
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.2.0
MarkupSafe==2.0.1
matplotlib==3.3.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
numpy==1.19.1
oauthlib==3.1.1
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.5.0
plotly==5.6.0
preprocessor==1.1.3
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.2
pyzmq==22.3.0
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.2
scipy==1.7.3
self==0.0.0
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.31
stack-data==0.1.4
streamlit==1.5.1
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
textblob==0.17.1
threadpoolctl==3.0.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tweepy==4.4.0
typing_extensions==4.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
users==1.0.dev0
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wordcloud==1.8.1
zipp==3.7.0
